infraipallocation: move `GetCiliumHostIPsFromFile`
This commit moves the function `node.GetCiliumHostIPsFromFile` from
package `pkg/node` into package `daemon/infraendpoints/` where it's
used by the `infraIPAllocator`.
